# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Application API for project WineApp",
        "version": "v1"
    },
    "paths": {
...
        "/api/v1/wineRecord": {
            "post": {
                "tags": [
                    "WineRecord"
                ],
                "summary": "Create WineRecord",
                "description": "Create WineRecord",
                "operationId": "createWineRecord",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "wine_evidence_id",
                                    "wine_record_type_id",
                                    "date"
                                ],
                                "properties": {
                                    "wine_evidence_id": {
                                        "type": "integer"
                                    },
                                    "wine_record_type_id": {
                                        "type": "integer"
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "is_in_progress": {
                                        "type": "boolean"
                                    },
                                    "date_to": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "title": {
                                        "type": "string"
                                    },
                                    "data": {
                                        "type": "string"
                                    },
                                    "note": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Response Successfull",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/v1/wineRecord/{wineRecordId}": {
            "get": {
                "tags": [
                    "WineRecord"
                ],
                "summary": "Show wineRecord by wineRecordId",
                "description": "Show wineRecord by wineRecordId",
                "operationId": "showWineRecordById",
                "parameters": [
                    {
                        "name": "wineRecordId",
                        "in": "path",
                        "description": "WineRecord ID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Response Successfull",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "put": {
                "tags": [
                    "WineRecord"
                ],
                "summary": "Update WineRecord",
                "description": "Update WineRecord by selected wineRecordId",
                "operationId": "updateWineRecord",
                "parameters": [
                    {
                        "name": "wineRecordId",
                        "in": "path",
                        "description": "WineRecord ID",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "wine_evidence_id",
                                    "wine_record_type_id",
                                    "date"
                                ],
                                "properties": {
                                    "wine_evidence_id": {
                                        "type": "integer"
                                    },
                                    "wine_record_type_id": {
                                        "type": "integer"
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "is_in_progress": {
                                        "type": "boolean"
                                    },
                                    "date_to": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "title": {
                                        "type": "string"
                                    },
                                    "data": {
                                        "type": "string"
                                    },
                                    "note": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Response Successfull",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/v1/wineRecord/wineEvidence/{wineEvidenceId}": {
            "get": {
                "tags": [
                    "WineRecord"
                ],
                "summary": "Show wineRecord list by wineEvidenceId",
                "description": "Show wineRecord list by wineEvidenceId",
                "operationId": "showWineRecordByWineEvidenceId",
                "parameters": [
                    {
                        "name": "wineEvidenceId",
                        "in": "path",
                        "description": "WineEvidence ID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Response Successfull",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/v1/wineRecordType": {
            "get": {
                "tags": [
                    "WineRecordType"
                ],
                "summary": "Show all WineRecordTypes",
                "description": "Show all WineRecordTypes",
                "operationId": "showWineRecordType",
                "responses": {
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "post": {
                "tags": [
                    "WineRecordType"
                ],
                "summary": "Create WineRecordType",
                "description": "Create WineRecordType",
                "operationId": "createWineRecordType",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "color": {
                                        "type": "string"
                                    },
                                    "note": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Response Successfull",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
    }
}